+++ src/Physics/Material/Material_Models/Zero_Energy_Control/global_control.jl
@@ -14,7 +14,7 @@ function control_name()
     return "Global"
 end
 
-function compute_control(datamanager::Module, nodes::Union{SubArray,Vector{Int64}}, material_parameter, time::Float32, dt::Float32)
+function compute_control(datamanager::Module, nodes::Union{SubArray,Vector{Int64}}, material_parameter, time::Float64, dt::Float64)
     rotation::Bool = false
     if "Angles" in datamanager.get_all_field_keys()
         rotation = true
@@ -24,11 +24,11 @@ function compute_control(datamanager::Module, nodes::Union{SubArray,Vector{Int64
     end
     dof = datamanager.get_dof()
     defGrad = datamanager.get_field("Deformation Gradient")
-    bond_force = datamanager.create_constant_bond_field("Bond Forces", Float32, dof)
+    bond_force = datamanager.create_constant_bond_field("Bond Forces", Float64, dof)
     bondGeom = datamanager.get_field("Deformed Bond Geometry", "N")
     bondGeomNP1 = datamanager.get_field("Deformed Bond Geometry", "NP1")
     Kinv = datamanager.get_field("Inverse Shape Tensor")
-    zStiff = datamanager.create_constant_node_field("Zero Energy Stiffness", Float32, "Matrix", dof)
+    zStiff = datamanager.create_constant_node_field("Zero Energy Stiffness", Float64, "Matrix", dof)
     CVoigt = get_Hooke_matrix(material_parameter, material_parameter["Symmetry"], dof)
     zStiff = create_zero_energy_mode_stiffness(nodes, dof, CVoigt, angles, Kinv, zStiff, rotation)
     bond_force = get_zero_energy_mode_force(nodes, zStiff, defGrad, bondGeom, bondGeomNP1, bond_force)
